+++ src/TwigComponent/tests/Fixtures/Component/Table.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Symfony\UX\TwigComponent\Tests\Fixtures\Component;
+
+use Symfony\UX\TwigComponent\Attribute\AsTwigComponent;
+
+#[AsTwigComponent('table')]
+final class Table
+{
+    public ?string $caption = null;
+    public array $headers;
+    public array $data;
+}